build: provide importable .tar.gz files for greater accessibility
.tar.gz files are now checked into version control under
the ./dist/ folder.

We update the README and Makefile accordingly. We also clear
some cruft out of the top of the README in order to make it
more useful to folks who are inbound from the upcoming Olive
release blog post.

Currently, there is no automation or CI checks to keep
the ./dist/ foldre up-to-date with the OLX.
This will be fixed: https://github.com/openedx/openedx-test-course/issues/7

Co-Authored-By: Sarina Canelake <sarina@tcril.org>